Refactor movement inside a Move object, Update files accordingly.
This one might not be necessary, and it's definitly more headache now then it will be worth, but my plan is to refactor all game rules from board into a new object, thus all the functions that caused me a headache are not going to be there anyway. I also plan to add a makeSpecialMove method to board, which will be able to take an array of moves. This seems to make more sense then just passing in an array of arrays of squares.