Add Promotion of Pawns
This involved both refactoring the SpecialMoves to include insertions instead of promotions, which would facilitate better stuff like bughouse, and reversing the check order to first check if a move is a special move. All in all stuff that makes sense. Game is officially playable though still doesn't detect stalemate or other kinds of draw...