Now a multiplayer game, quite a large overhaul.
not all changes will be recorded here, but the socket and server systems have been changed and, to fit, a lot of the logic of the game have been shuffled around (also to hopefully make more sense). Some things are still quite a mess, and the castling, stalemate and promotion problems will probably not be easily fixable in the current version. But at least everything is working for the most part.